move utils/plural and utils/date to utils/format
